move the main class into a dedicated file, no more mixing of defining using and creating classes
